refactor(ssa-workflow): split stochastic input iterators
